Removed the need for some plugins with the aim to eliminate them completely
Numberous modifications to the site structure including moving _pages to pages so a plugin
isn't needed to describe the pages.
Created a Liquid template for describing Series and displaying a Series post list so the
series plugin could be deleted.
Added SASS capabilities to the site (although not currently making proper use of it other
than for css minification).
Removed the need for Grunt to process files before processing by adding minified verions
of Javascript files and calling each individually in the header/footer files.