Updated site navigation to be recursive
The crappy navigation script I had was flawed, I was only able to get to 2 levels of sub-menus
because it was all hard coded. I've replaced that with a more obvious, recursive algorithm that
just includes the nav.html file if there are sub-menus!
Whilst I was at it, I also incorporated the mini-social.html and social.html into the default.html
layout file. I'm trying to reduce the number of files in the template ready to create my own proper
Jekyll theme that I can include for each of my sites and only have to update from a single location.