Updated all flickr_image and flickr_photoset tags
It seems there is a problem with the flickr plugin I have been using for years as there are consistent build errors when using it. For that reason I have decided to remove all the tags and replace with the previously generated HTML. I may clean this up in future, but probably not. It is not a priority, but maybe getting back to writing will be. We'll see.